Core - added XUR8 write support for vector properties. Fixed logic bug in DATA8 section with an errorneous cast to byte for object property masks, switched to a uint.

+++ XUIHelper.Core/XUIHelper.Core/XUR/V8/Sections/DATA8Section.cs
@@ -528,7 +528,7 @@ namespace XUIHelper.Core
                         {
                             if (propertyDefinition == property.PropertyDefinition)
                             {
-                                thisPropertyMask |= (byte)(1 << propertyDefinitionIndex);
+                                thisPropertyMask |= (uint)(1 << propertyDefinitionIndex);
                                 break;
                             }
                         }
